requirements.txt
Django==5.0
djangorestframework==3.14.0
python-telegram-bot==22.6
django-cors-headers==4.3.1
drf-spectacular==0.27.0
Pillow==10.1.0
python-dotenv==1.0.0
asgiref==3.7.2
torch>=2.0.0
transformers>=4.35.0
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
